+++ scripts/init.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-rm -rf node_modules
-rm -rf vendor
-npm install && node_modules/.bin/bower install