+++ scripts/server.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-rm -rf public
-node_modules/.bin/brunch watch --server