+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.5.0 (30 April 2014)
+* **breaking change** IE8 is no longer supported by angular, this seed has been updated accordingly. Support is now IE9+
+* New versions of angular, jquery & other libs
+* Protractor is now used for e2e tests
+* Npm scripts are now used instead of .bat & .sh scripts, check the README.md or package.json files for the list of commands
+* Lodash has been added to the vendor files
+
 # 0.4.1 (16 Mars 2014)
 * New version of angular, last one supporting IE8, new releases will only support IE9+
 * Node-sass has been updated and we can use the official twitter bootstrap sass port from now on